refactor: combine Node and NodeData
